BREAKING CHAGE: re-structure VOS website live build system (#23)
