--- composer.json
@@ -1,6 +1,0 @@
-{
-    "require": {
-        "vlucas/phpdotenv": "^5.6",
-        "guzzlehttp/guzzle": "^7.8"
-    }
-}
