--- data/staff.json
@@ -1,172 +1,0 @@
-{
-  "vot4": [
-    {
-      "id": 9623142,
-      "role": "Host, Mappooler, GFX / VFX, Mapper, Playtester, Referee, Commentator"
-    },
-    {
-      "id": 26012543,
-      "role": "Mappooler, Referee"
-    },
-    {
-      "id": 18397349,
-      "role": "Mappooler"
-    },
-    {
-      "id": 11833538,
-      "role": "Mappooler, Mapper"
-    },
-    {
-      "id": 16039831,
-      "role": "GFX / VFX, Mapper, Playtester, Commentator"
-    },
-    {
-      "id": 14083855,
-      "role": "GFX / VFX"
-    },
-    {
-      "id": 13981991,
-      "role": "GFX / VFX"
-    },
-    {
-      "id": 9912966,
-      "role": "GFX / VFX"
-    },
-    {
-      "id": 14001000,
-      "role": "GFX / VFX"
-    },
-    {
-      "id": 17302272,
-      "role": "Mapper"
-    },
-    {
-      "id": 7169208,
-      "role": "Mapper"
-    },
-    {
-      "id": 22522738,
-      "role": "Mapper"
-    },
-    {
-      "id": 22515524,
-      "role": "Mapper, Referee"
-    },
-    {
-      "id": 3724819,
-      "role": "Mapper"
-    },
-    {
-      "id": 8631719,
-      "role": "Mapper"
-    },
-    {
-      "id": 2865172,
-      "role": "Mapper"
-    },
-    {
-      "id": 3738344,
-      "role": "Mapper"
-    },
-    {
-      "id": 14520910,
-      "role": "Mapper"
-    },
-    {
-      "id": 17916791,
-      "role": "Mapper"
-    },
-    {
-      "id": 12510704,
-      "role": "Mapper"
-    },
-    {
-      "id": 2345156,
-      "role": "Mapper"
-    },
-    {
-      "id": 11056763,
-      "role": "Mapper"
-    },
-    {
-      "id": 1109122,
-      "role": "Mapper"
-    },
-    {
-      "id": 26190106,
-      "role": "Mapper"
-    },
-    {
-      "id": 2228401,
-      "role": "Playtester"
-    },
-    {
-      "id": 11411697,
-      "role": "Playtester"
-    },
-    {
-      "id": 13630137,
-      "role": "Playtester"
-    },
-    {
-      "id": 21290592,
-      "role": "Referee, Commentator"
-    },
-    {
-      "id": 19207842,
-      "role": "Referee"
-    },
-    {
-      "id": 17148657,
-      "role": "Referee"
-    },
-    {
-      "id": 7109317,
-      "role": "Referee"
-    },
-    {
-      "id": 10129901,
-      "role": "Referee"
-    },
-    {
-      "id": 829469,
-      "role": "Referee"
-    },
-    {
-      "id": 13456818,
-      "role": "Streamer, Statistician"
-    },
-    {
-      "id": 7789926,
-      "role": "Streamer"
-    },
-    {
-      "id": 15815791,
-      "role": "Streamer"
-    },
-    {
-      "id": 19817503,
-      "role": "Streamer"
-    },
-    {
-      "id": 22069182,
-      "role": "Streamer, Commentator"
-    },
-    {
-      "id": 20569104,
-      "role": "Streamer"
-    },
-    {
-      "id": 18916920,
-      "role": "Streamer"
-    },
-    {
-      "id": 24042710,
-      "role": "Commentator"
-    },
-    {
-      "id": 10494860,
-      "role": "Statistician"
-    }
-  ]
-}
