+++ src/private/controller/VotController.php
@@ -1,0 +1,6 @@
+<?php
+# Not so much like static types, but at least it does feel better having this here
+declare(strict_types=1);
+
+// TODO: Make this like a General File Manager (GFM). Therefore, proper handling needed
+include_once __DIR__ . '/../view/Navigation.php';
