+++ src/private/resource/Staff.json
@@ -1,0 +1,246 @@
+{
+  "vot3": [
+    {
+      "id": 9623142,
+      "role": "Host, Mappooler, Custom Mapper, Playtester, Referee, Commentator"
+    },
+    {
+      "id": 16039831,
+      "role": "Host, Mappooler, Designer, Custom Mapper, Playtester, Commentator"
+    },
+    {
+      "id": 16688499,
+      "role": "Host, Mappooler, Designer, Custom Mapper, Playtester, Referee, Streamer, Commentator"
+    },
+    {
+      "id": 7169208,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 22522738,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 22515524,
+      "role": "Custom Mapper, Playtester, Referee"
+    },
+    {
+      "id": 8301957,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 6291741,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 10547864,
+      "role": "Composer"
+    },
+    {
+      "id": 11411697,
+      "role": "Playtester"
+    },
+    {
+      "id": 26012543,
+      "role": "Referee, Commentator"
+    },
+    {
+      "id": 17148657,
+      "role": "Referee"
+    },
+    {
+      "id": 19787040,
+      "role": "Referee"
+    },
+    {
+      "id": 13456818,
+      "role": "Referee"
+    },
+    {
+      "id": 7789926,
+      "role": "Streamer"
+    },
+    {
+      "id": 15815791,
+      "role": "Streamer"
+    },
+    {
+      "id": 21290592,
+      "role": "Commentator"
+    },
+    {
+      "id": 10494860,
+      "role": "Commentator"
+    }
+  ],
+  "vot4": [
+    {
+      "id": 9623142,
+      "role": "Host, Mappooler, GFX / VFX, Custom Mapper, Playtester, Referee, Commentator"
+    },
+    {
+      "id": 26012543,
+      "role": "Mappooler, Referee"
+    },
+    {
+      "id": 18397349,
+      "role": "Mappooler"
+    },
+    {
+      "id": 11833538,
+      "role": "Mappooler, Custom Mapper"
+    },
+    {
+      "id": 16039831,
+      "role": "GFX / VFX, Custom Mapper, Playtester, Commentator"
+    },
+    {
+      "id": 14083855,
+      "role": "GFX / VFX"
+    },
+    {
+      "id": 13981991,
+      "role": "GFX / VFX"
+    },
+    {
+      "id": 9912966,
+      "role": "GFX / VFX"
+    },
+    {
+      "id": 14001000,
+      "role": "GFX / VFX"
+    },
+    {
+      "id": 17302272,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 7169208,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 22522738,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 22515524,
+      "role": "Custom Mapper, Referee"
+    },
+    {
+      "id": 3724819,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 8631719,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 2865172,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 3738344,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 14520910,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 17916791,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 12510704,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 2345156,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 11056763,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 1109122,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 26190106,
+      "role": "Custom Mapper"
+    },
+    {
+      "id": 2228401,
+      "role": "Playtester"
+    },
+    {
+      "id": 11411697,
+      "role": "Playtester"
+    },
+    {
+      "id": 13630137,
+      "role": "Playtester"
+    },
+    {
+      "id": 21290592,
+      "role": "Referee, Commentator"
+    },
+    {
+      "id": 19207842,
+      "role": "Referee"
+    },
+    {
+      "id": 17148657,
+      "role": "Referee"
+    },
+    {
+      "id": 7109317,
+      "role": "Referee"
+    },
+    {
+      "id": 10129901,
+      "role": "Referee"
+    },
+    {
+      "id": 829469,
+      "role": "Referee"
+    },
+    {
+      "id": 13456818,
+      "role": "Streamer, Statistician"
+    },
+    {
+      "id": 7789926,
+      "role": "Streamer"
+    },
+    {
+      "id": 15815791,
+      "role": "Streamer"
+    },
+    {
+      "id": 19817503,
+      "role": "Streamer"
+    },
+    {
+      "id": 22069182,
+      "role": "Streamer, Commentator"
+    },
+    {
+      "id": 20569104,
+      "role": "Streamer"
+    },
+    {
+      "id": 18916920,
+      "role": "Streamer"
+    },
+    {
+      "id": 24042710,
+      "role": "Commentator"
+    },
+    {
+      "id": 10494860,
+      "role": "Statistician"
+    }
+  ]
+}
