+++ src/public/admin/AccessToken.php
@@ -1,0 +1,8 @@
+<?php
+require_once '../../private/controller/api/OsuUserAccessToken.php';
+?>
+
+<!--
+<h1 style="display: flex; justify-content: center; align-items: center;">User access token received successfully. You may use it legally now xD</h1>
+<p style="display: flex; justify-content: end; align-items: end; font-weight: bold;"><a href="/../user/home.php">Home Page</a></p>
+-->
