+++ src/public/admin/LogOutInterface.php
@@ -1,0 +1,8 @@
+<?php
+require_once '../../private/controller/user/LogOut.php';
+?>
+
+<!--
+<h1 style="display: flex; justify-content: center; align-items: center;">User logout successfully!!</h1>
+<p style="display: flex; justify-content: end; align-items: end; font-weight: bold;"><a href="/../user/home.php">Home Page</a></p>
+-->
