+++ src/public/assets/ico/favicon-32x32.png
