+++ src/public/assets/ico/favicon.ico
