+++ src/public/index.php
@@ -1,0 +1,11 @@
+<?php
+# Not so much like static types, but at least it does feel better having this here
+declare(strict_types=1);
+?>
+
+<h1 style="display: flex; justify-content: center; align-items: center; color: coral;">Hello there, your website is working!</h1>
+<p style="display: flex; justify-content: end; align-items: end; font-weight: bold;"><a href="user/Home.php">Home Page</a></p>
+
+<?php
+phpinfo();
+?>
