+++ src/public/user/Home.php
@@ -1,0 +1,58 @@
+<?php
+# Not so much like static types, but at least it does feel better having this here
+declare(strict_types=1);
+
+require_once '../../private/controller/VotController.php';
+?>
+
+<section class="tournament-news">
+    <div class="announcement-box" id="latest-news">
+        <h2>VIETNAMESE OSU!TAIKO TOURNAMENT</h2>
+        <p>Welcome to the website of the biggest vietnamese osu!taiko tournament battlefield ! In here, you can look at everything related to VOT, with pages for all our current and past tournaments!</p>
+
+        <div class="youtube-iframe-container">
+            <iframe width="560" height="315" src="https://www.youtube.com/embed/Y1-gjfcoJnU?autoplay=1&amp;showinfo=0&amp;controls=0&amp;autohide=1&amp;mute=1&amp;loop=1" title="VOT4" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+        </div>
+    </div>
+
+    <div class="announcement-box">
+        <!-- TODO: sent query parameters with this redirect to indicate it's the Landing page of which tournament? -->
+        <a href="Landing.php">
+            <h2>VIETNAMESE OSU!TAIKO TOURNAMENT 4</h2>
+            <img src="/../assets/img/VOT4.webp" alt="VOT4 Banner">
+            <p>
+                The Vietnamese osu!taiko Tournament, a cornerstone event in the Vietnamese osu!taiko community. It is our 1v1 tournament for Vietnamese players with no rank restrictions, open to all skill levels. The players will first compete in Qualifiers. Then, they will progress through a Double-elimination Bracket, keeping the competitive syndrome burning until the very last crowning glory - the Grand Finals! Are you ready to make history in the Vietnamese osu!taiko scene? VOT4 is calling, and it's time to answer! Let the drums roll and may the best taiko player win!
+            </p>
+        </a>
+    </div>
+
+    <div class="announcement-box">
+        <a href="#">
+            <h2>VIETNAMESE OSU!TAIKO TOURNAMENT 3</h2>
+            <img src="/../assets/img/VOT3.webp" alt="VOT3 Banner">
+            <p>
+                The Vietnamese osu!taiko Tournament, a cornerstone event in the Vietnamese osu!taiko community. It is our 1v1 tournament for Vietnamese players with no rank restrictions, open to all skill levels. The players will first compete in Qualifiers. Then, they will progress through a Double-elimination Bracket, keeping the competitive syndrome burning until the very last crowning glory - the Grand Finals! Are you ready to make history in the Vietnamese osu!taiko scene? VOT3 is calling, and it's time to answer! Let the drums roll and may the best taiko player win!
+            </p>
+        </a>
+    </div>
+
+    <div class="announcement-box">
+        <a href="#">
+            <h2>VIETNAMESE OSU!TAIKO TOURNAMENT 2</h2>
+            <img src="/../assets/img/VOT2.webp" alt="VOT2 Banner">
+            <p>
+                The Vietnamese osu!taiko Tournament, a cornerstone event in the Vietnamese osu!taiko community. It is our 1v1 tournament for Vietnamese players with no rank restrictions, open to all skill levels. The players will first compete in Qualifiers. Then, they will progress through a Double-elimination Bracket, keeping the competitive syndrome burning until the very last crowning glory - the Grand Finals! Are you ready to make history in the Vietnamese osu!taiko scene? VOT2 is calling, and it's time to answer! Let the drums roll and may the best taiko player win!
+            </p>
+        </a>
+    </div>
+
+    <div class="announcement-box">
+        <a href="#">
+            <h2>VIETNAMESE OSU!TAIKO TOURNAMENT 1</h2>
+            <img src="/../assets/img/VOT1.webp" alt="VOT1 Banner">
+            <p>
+                The Vietnamese osu!taiko Tournament, a cornerstone event in the Vietnamese osu!taiko community. It is our 1v1 tournament for Vietnamese players with no rank restrictions, open to all skill levels. The players will first compete in Qualifiers. Then, they will progress through a Double-elimination Bracket, keeping the competitive syndrome burning until the very last crowning glory - the Grand Finals! Are you ready to make history in the Vietnamese osu!taiko scene? VOT1 is calling, and it's time to answer! Let the drums roll and may the best taiko player win!
+            </p>
+        </a>
+    </div>
+</section>
